+++ tests/Mailtrap.IntegrationTests/TestConstants/UrlSegmentsTestConstants.cs
@@ -28,4 +28,6 @@ internal static class UrlSegmentsTestConstants
     internal static string BatchEmailSegment { get; } = "batch";
     internal static string StatsSegment { get; } = "stats";
     internal static string ApiTokensSegment { get; } = "api_tokens";
+    internal static string OrganizationsSegment { get; } = "organizations";
+    internal static string SubAccountsSegment { get; } = "sub_accounts";
 }