+++ tests/Mailtrap.UnitTests/Organizations/OrganizationSubAccountCollectionResourceTests.cs
@@ -1,0 +1,47 @@
+namespace Mailtrap.UnitTests.Organizations;
+
+
+[TestFixture]
+internal sealed class OrganizationSubAccountCollectionResourceTests
+{
+    private readonly IRestResourceCommandFactory _commandFactoryMock = Mock.Of<IRestResourceCommandFactory>();
+    private readonly Uri _resourceUri = EndpointsTestConstants.ApiDefaultUrl
+        .Append(
+            UrlSegmentsTestConstants.ApiRootSegment,
+            UrlSegmentsTestConstants.OrganizationsSegment)
+        .Append(TestContext.CurrentContext.Random.NextLong())
+        .Append(UrlSegmentsTestConstants.SubAccountsSegment);
+
+
+    #region Constructor
+
+    [Test]
+    public void Constructor_ShouldThrowArgumentNullException_WhenCommandFactoryIsNull()
+    {
+        var act = () => new OrganizationSubAccountCollectionResource(null!, _resourceUri);
+
+        act.Should().Throw<ArgumentNullException>();
+    }
+
+    [Test]
+    public void Constructor_ShouldThrowArgumentNullException_WhenUriIsNull()
+    {
+        var act = () => new OrganizationSubAccountCollectionResource(_commandFactoryMock, null!);
+
+        act.Should().Throw<ArgumentNullException>();
+    }
+
+    [Test]
+    public void ResourceUri_ShouldBeInitializedProperly()
+    {
+        var client = CreateResource();
+
+        client.ResourceUri.Should().Be(_resourceUri);
+    }
+
+    #endregion
+
+
+
+    private OrganizationSubAccountCollectionResource CreateResource() => new(_commandFactoryMock, _resourceUri);
+}